+++ OUDS/Core/Tokens/RawTokens/Sources/TypeAliases/TypographyRawTokens+Aliases.swift
@@ -16,8 +16,8 @@ import Foundation
 /// In the global design system tool, *font family* raw tokens are basically `String` values, to keep grammar clean and clear with design system grammar.
 public typealias TypographyFontFamilyRawToken = String
 
-/// In the global design system tool, *font weight* raw tokens are basically `String` values, to keep grammar clean and clear with design system grammar.
-public typealias TypographyFontWeightRawToken = String
+/// In the global design system tool, *font weight* raw tokens are basically `Int` values, to keep grammar clean and clear with design system grammar.
+public typealias TypographyFontWeightRawToken = Int
 
 /// In the global design system tool, *font size* raw tokens are basically `CGFloat` values, to keep grammar clean and clear with design system grammar.
 public typealias TypographyFontSizeRawToken = CGFloat